[hwajun] refactor : repository 폴더 구조 변경
