+++ src/main/java/com/ohgoodteam/ohgoodpay/shorts/controller/feed/ShortsFeedController.java
@@ -111,7 +111,7 @@ public class ShortsFeedController {
 
     }
 
-
+    // 포인트 적립 요청
     @PostMapping("/point/earn")
     public ResponseEntity<ShortsPointEarnResponseDto> earnPoint(
         @RequestBody ShortsPointEarnRequestDto requestDto
@@ -188,14 +188,11 @@ public class ShortsFeedController {
         }
     }
 
-
-
+    // 특정 쇼츠 조회
     @GetMapping("/{shortsId}")
     public ResponseEntity<ShortsFeedDataDto> getSpecificShorts(
         @PathVariable (value = "shortsId") Long shortsId
     ){
         return ResponseEntity.ok(shortsFeedService.getSpecificShorts(shortsId));
     }
-
-
 }
