+++ src/main/java/com/ohgoodteam/ohgoodpay/shorts/repository/SubscriptionRepository.java
@@ -1,10 +1,11 @@
-package com.ohgoodteam.ohgoodpay.shorts.repository.mypage;
+package com.ohgoodteam.ohgoodpay.shorts.repository;
 
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import com.ohgoodteam.ohgoodpay.common.entity.SubscriptionEntity;
+import com.ohgoodteam.ohgoodpay.shorts.repository.mypage.SubscriptionRepositoryCustom;
 
 public interface SubscriptionRepository extends JpaRepository<SubscriptionEntity, Long>, SubscriptionRepositoryCustom {
