+++ src/main/java/com/ohgoodteam/ohgoodpay/shorts/service/feed/ShortsFeedServiceImpl.java
@@ -17,6 +17,7 @@ import com.ohgoodteam.ohgoodpay.shorts.dto.response.feed.*;
 import com.ohgoodteam.ohgoodpay.shorts.repository.CommentRepository;
 import com.ohgoodteam.ohgoodpay.shorts.repository.ReactionRepository;
 import com.ohgoodteam.ohgoodpay.shorts.repository.ShortsRepository;
+
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
